refactor: Create a new component to display OwnerRecipient
In WhoHasAccess, it was difficult to understand that part of the code
was just to display the owner of the file.Now, this code is
in a new component dedicatedto display an owner recipient.
