+++ packages/cozy-sharing/src/components/WhoHasAccess.jsx
@@ -1,14 +1,13 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { useClient, useQuery, hasQueryBeenLoaded } from 'cozy-client'
 import Recipient from './Recipient/Recipient'
 import LinkRecipient from './Recipient/LinkRecipient'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import List from 'cozy-ui/transpiled/react/MuiCozyTheme/List'
 
-import { buildInstanceSettingsQuery } from '../queries/queries'
 import { filterAndReworkRecipients } from '../helpers/recipients'
 import { usePrevious } from '../helpers/hooks'
+import OwnerRecipient from './Recipient/OwnerRecipient'
 
 /**
  * Displays a warning if some contacts are waiting for confirmation of their sharing
@@ -42,14 +41,6 @@ const WhoHasAccess = ({
   onUpdateShareLinkPermissions,
   onRevokeLink
 }) => {
-  const client = useClient()
-
-  const instanceSettingsQuery = buildInstanceSettingsQuery()
-  const instanceSettingsResult = useQuery(
-    instanceSettingsQuery.definition,
-    instanceSettingsQuery.options
-  )
-
   const previousLink = usePrevious(link)
   const previousRecipients = usePrevious(recipients)
 
@@ -59,8 +50,6 @@ const WhoHasAccess = ({
     recipients,
     previousRecipients
   )
-
-  const public_name = instanceSettingsResult?.data?.attributes?.public_name
 
   return (
     <div className={className}>
@@ -82,14 +71,7 @@ const WhoHasAccess = ({
           />
         )}
 
-        {hasQueryBeenLoaded(instanceSettingsResult) && (
-          <Recipient
-            isOwner={true}
-            status="owner"
-            instance={client.options.uri}
-            public_name={public_name}
-          />
-        )}
+        <OwnerRecipient />
 
         {recipientsToDisplay.map(recipient => {
           const recipientConfirmationData = recipientsToBeConfirmed.find(
